promote print_number input to long int

+++ main.h
@@ -4,7 +4,7 @@
 
 int _printf(const char *format, ...);
 int _putchar(char c);
-int print_number(int n, const char **format, int fc);
+int print_number(long int n, const char **format, int fc);
 int print_str(char *str);
 int print_binary(unsigned int n, const char **format, int fc);
 int flag_count(const char *str);
